Rename Inspect helper method to Dump, and significantly improve its output:
Print type of the object itself at the beginning.

Property values are printed in such a way that they can be directly copy/pasted as input to csx.eval; e.g. strings are quoted.

For non-primitive objects and collections, print the actual type of the object in addition to property type.

For enumerables, print up to 1000 elements.